requirements.txt
fastapi
uvicorn
python-multipart
ffmpeg-python
torch
openai-whisper
